Fix: heredoc 버그 수정
heredoc 일때 heredoc 종료 문자열 저장,  리드라인 받고 만들어진 임시
파일을 이용해서 리다이렉트로 처리하여 넘긴다.해당 부분 처리하면서 인덱스
값이 늘어나지 않아서 나던 세그폴트와, 임시파일을 argv에여러번 저장 하던
버그를 한번만 저장하게 수정.

+++ srcs/list/cmd_list.c
@@ -3,14 +3,27 @@
 /*                                                        :::      ::::::::   */
 /*   cmd_list.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jinyoo <jinyoo@student.42seoul.kr>         +#+  +:+       +#+        */
+/*   By: minjupar <minjupar@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/23 01:16:57 by minjupar          #+#    #+#             */
-/*   Updated: 2022/04/27 15:37:31 by jinyoo           ###   ########.fr       */
+/*   Updated: 2022/04/28 00:37:24 by minjupar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+int	here_filename(t_cmd *head)
+{
+	int	result;
+
+	result = 0;
+	while (head)
+	{
+		head = head->next;
+		result ++;
+	}
+	return (result);
+}
 
 t_cmd	*create_cmd_node(t_cmd *prev)
 {
@@ -26,6 +39,7 @@ t_cmd	*create_cmd_node(t_cmd *prev)
 	node->argc = 0;
 	node->input = NULL;
 	node->output = NULL;
+	node->heredoc = NULL;
 	node->here_filename = NULL;
 	node->next = NULL;
 	node->prev = prev;
@@ -60,7 +74,8 @@ void	delete_cmd_list(t_cmd **cmd)
 	temp = *cmd;
 	while (temp)
 	{
-		unlink(temp->here_filename);
+		if (temp->heredoc)
+			unlink(temp->here_filename);
 		del_node = temp;
 		temp = temp->next;
 		ft_free_two_ptr(del_node->argv);
